Changed AdminService to WorkshopCodeService and admin package to workshopCode, added GrantUserAdminRequest to User, fixed error messages in StationController.
