+++ ClimateChecker/src/main/java/Ontdekstation013/ClimateChecker/features/station/endpoint/StationController.java
@@ -1,7 +1,7 @@
 package Ontdekstation013.ClimateChecker.features.station.endpoint;
 
 import Ontdekstation013.ClimateChecker.exception.InvalidArgumentException;
-import Ontdekstation013.ClimateChecker.features.admin.AdminService;
+import Ontdekstation013.ClimateChecker.features.workshopCode.WorkshopCodeService;
 import Ontdekstation013.ClimateChecker.features.measurement.MeasurementService;
 import Ontdekstation013.ClimateChecker.features.station.Station;
 import Ontdekstation013.ClimateChecker.features.station.StationService;
@@ -24,12 +24,12 @@ import java.time.format.DateTimeFormatter;
 public class StationController {
     private final StationService stationService;
     private final MeasurementService measurementService;
-    private final AdminService adminService;
+    private final WorkshopCodeService adminService;
 
-    public StationController(StationService stationService, MeasurementService measurementService, AdminService adminService){
+    public StationController(StationService stationService, MeasurementService measurementService, WorkshopCodeService workshopCodeService){
         this.stationService = stationService;
         this.measurementService = measurementService;
-        this.adminService = adminService;
+        this.adminService = workshopCodeService;
     }
 
     @GetMapping("{id}")
